requirements.txt
python-telegram-bot>=20.0
solana>=0.29.0
base58>=2.0.0
cryptography>=40.0.0
python-dotenv>=0.19.0
aiohttp>=3.8.0
psutil>=5.8.0
